% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_correction.R
\name{get_alphas}
\alias{get_alphas}
\title{Calculate Benjamini-Hochberg alphas for a list of p-values.}
\usage{
get_alphas(p_values, Q = 0.05)
}
\arguments{
\item{p_values}{A list of p-values.}

\item{Q}{the false discovery rate. If not provided, a default value of 0.05
is used.}
}
\value{
A list of alpha values or an error message if the provided p-values
are not valid.
}
\description{
This function calculates alphas for a list of p-values and for a given false
discovery rate (Q). If Q is not provided, a default value of 0.05 is used.
The Benjamini-Hochberg alpha corrected is calculated as:
alpha=(i/m)Q, where:
i = the individual p-value's rank in the list of p-values,
m = the total number of tests, and
Q = the false discovery rate.
The function does not expect the p-values to be sorted but it will return
alpha values for the sorted p-values
so you are strongly encouraged to sort the p-values before passing them to
this function - this will make the returned alpha values more readable.
}
